ifcopenshell-rules/requirements.txt
click>=8.0.0
pyyaml>=6.0
ifcopenshell
numpy
shapely
pydantic>=2.0
